# cuda4dnn(region): optimize kernels