--- a/modules/dnn/src/cuda4dnn/kernels/region.hpp
+++ b/modules/dnn/src/cuda4dnn/kernels/region.hpp
@@ -13,19 +13,12 @@
 namespace cv { namespace dnn { namespace cuda4dnn { namespace kernels {
 
     template <class T>
-    void sigmoid_strided(const csl::Stream& stream, csl::Span<T> output, csl::View<T> input, std::size_t n, std::size_t stride, std::size_t offset);
-
-    template <class T>
-    void softmax_strided(const csl::Stream& stream, csl::Span<T> output, csl::View<T> input, std::size_t n, std::size_t stride, std::size_t offset);
-
-    template <class T>
-    void region_finalize(const csl::Stream& stream, csl::Span<T> output, csl::View<T> input, csl::View<T> bias,
+    void region(const csl::Stream& stream, csl::Span<T> output, csl::View<T> input, csl::View<T> bias,
         T object_prob_cutoff, T class_prob_cutoff,
-        std::size_t height_norm, std::size_t width_norm,
+        std::size_t boxes_per_cell, std::size_t box_size,
         std::size_t rows, std::size_t cols,
-        std::size_t boxes_per_cell,
-        std::size_t box_size,
-        std::size_t classes);
+        std::size_t height_norm, std::size_t width_norm,
+        bool if_true_sigmoid_else_softmax);
 
 }}}} /* namespace cv::dnn::cuda4dnn::kernels */